#155 cleanup
